back-tasks-planok-main/requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers
psycopg2-binary
python-dotenv
langchain
langchain-openai
sqlparse==0.5.5